requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
PyJWT==2.8.0
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.1
thefuzz[speedup]==0.22.1
sentence-transformers>=3.0.1
torch
pillow
python-multipart

# Frontend map and static serving are handled via CDN (Leaflet, etc.)
